backend/requirements.txt
# PolaZhenjing Backend Dependencies
# FastAPI + Uvicorn
fastapi==0.115.6
uvicorn[standard]==0.34.0

# Database
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0

# Pydantic & Settings
pydantic==2.10.3
pydantic-settings==2.7.0

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.19

# HTTP Client (for AI providers)
httpx==0.28.1

# MkDocs (for static site generation)
mkdocs==1.6.1
mkdocs-material==9.5.47

# Testing
pytest==8.3.4
pytest-asyncio==0.24.0
httpx  # also used for test client

# Utilities
python-slugify==8.0.4
